# Improve performance of `brew info` by caching compiled regexes